benchmark: Reorganize json logic
We want all the json logic in one place as this will make it easier to
update between iterations. We put all json parameters into the json/default.json
file and upload each lambda's parameters with the corresponding lambda
to AWS. This will allow us to not have to change files whenever we want to
change a parameter.