split_spectra: Handle large files
We only have 512MB of disk space on lambda. Spectra files can be a
couple of gigabytes. As a result, we cannot store the entire file.
To handle this, we change split spectra to read in the file in parts
and create the spectra chunk files as we read in the file.