benchmark: Add timeout logic
For the combine and percolator steps, we check if the files are
present periodically to see if they're done. However, the files will
never appear if the lambda times out. To handle these cases, we have
the file check timeout if we haven't seen the file in a given period
of time.