benchmark: Improve output file checks
We only care about checking the target files, as the decoy files
are bad results. The results are not guaranteed to be the exact
same every time, as it matters on the order of results since
Percolator trains on a subset of the data. So we change the
test to check if the result of valid qvalues is within a given
range.

Also forgot to add the checks json.