benchmark: Generalizations + changes to EC2 logic
To make the comparison between EC2 and Lambda more fair, we upload
the files to S3, initiate the EC2 instance, and download the file
from S3.

To make it easier to implement this, we generalize the upload logic
and compute the necessary buckets and file names at the beginning so
the rest of the code can be more general.