benchmark: Add sorting to EC2
We want to compare how fast sorting is on EC2 versus Lambda. So
we add sorting logic to a python file that can be uploaded to EC2.

The EC2 instance we're using only supports Python 2, not Python 3.
So uploading util.py is bad since it requires boto3. To fix this,
we create a constant.py file that contains the regex constants.