benchmark: file_count function
This function counts the number of files for the run created in a
bucket. Since the number of output files is less deterministic now,
it seems easier to just count the number of files to determine the
number of lambdas. This also allows us to add in print statements
how many files have been created in each bucket so we can see that
progress is made.