lambda: Allow splits in analysis function
I tried running the benchmark on a COON file, and Tide segfaulted on
one of the instances. This segfault did not occur on my local computer.
The size of the input file, plus the size of the files Tide outputed,
plus the binary files were about the amount of allocated disk space.

It's been really annoying to have to manually adjust the batch sizes
to make split run fast enough and to make sure we don't run out of
space. So instead what we're going to do is optimize for making split
run fast. When a segfault occurs, we take the input file and split
it in half and run these files in new lambda instances.

This complicates the naming scheme a little bit, and I need to come
up with a better way to indicate that a split occurred, especially
if we have to do multiple rounds of splitting. The reason we need to
embed the splits in the name is AWS does not have a great way to
signal that all the files we want are present. We may have to
consider using a coordinator at some point, but for now, I added an
addition field to indicate that this file has been split and manually
increment the counter.