benchmark: Coordinator changes
We want to see how well Lambda scales with a coordinator. To do this,
we add a section for coordinator logic and we try to generalize some
of the Lambda logic so it can work for both Lambda and the coordinator
version.

This includes having the split logic just return offsets.