benchmark: Add logic for multiple benchmark instances
We want to see what happens if we send multiple requests in parallel.
To fetch logs for all requests, we cannot solely rely on timestamps
to indicate that the event occurred after the timestamp.

Since searching through logs only allows us to specify timestamps and
not request IDs, we will try to do this in two iterations. The first
iteration searches the log for a print statement we have added
to each Lambda. This will be the unique identifier for the thread.
The second iteration will take the log stream ID from the first
iteration to find the request information such as memory or
time duration.