benchmark: Update to handle new pipeline
Also from the benchmark, it's clear multiple combine functions can
be called from the same set. Choosing the last timestamp does not
seem to help as all are made about the same time. We're going to
take the last element and just wait for the rest to appear. If
there are extremely time differences between the elements, this
could be bad.