lambda: Print duration
Due to the request ID being reused, we can't trust the duration line
to figure out how long each specific request takes. Instead, we
calculate the duration. We only use this for the graphs. For the costs
we will still use the request line since we just need to sum the
durations.