ripple: Changes related to framework and formats
A lot of this still needs to be tested.

We add a confidence function that is faster than percolator to
help identify species. In front of the actual tide process, we
identify the specie using tide and the 1,000 most intense spectra.
Then we use the specie fasta file to analyze all spectra.