First round of changing offsets
We need to pass around a list of offsets, not the start, end bytes.
This started because we can't use a set header anymore. Some of the
header parameters affect Tide, so it's easier to just pass the header
offsets. Due to this, we changed the offsets logic to return an object
containing the offsets and header offsets.

In addition, for functions such as top, to prevent overloading memory,
we decide to pass in the array of offsets so we don't have to load in
all bytes at once.

I have only fixed combine, split and top so far.