ripple: Methyl DNA compression
Add logic for Methyl DNA compression. This pipeline compresses
Methyl DNA and decompresses afterwards.

One thing this pipeline brought to my attention is that sometimes
a function outputs multiple files that are needed for the next
stage of the pipeline. We don't want to trigger a lambda for all
of these files. To prevent this from happening, we add support for
suffix filtering. I use the suffix value to distinguish between
the different files that are saved. By filtering on just one of
those suffixes, we can prevent triggers for every file from occuring.