methyl: Update methyl application logic
It was too slow splitting the the methyl file when we selected every
newline as an offset. Instead we just return the last newline as an
offset for iterator since we don't need every newline for this format.
When I have time, I'll see if mzML format can be generalized to this
because this was the original format that caused me to return an offset
for every item.