mzML: Update offset logic
To save memory, we only return the start and end offset as we
don't need the rest. Also to simplify the logic, we implement
nextOffsets and don't call iterator nextOffsets. The mzML logic
is not the same as most of the other formats, so it makes sense
to have it's only nextOffsets function.