split_file: Start adding course / fine grain split support
The DNA compression application made me realize the split function
basically makes the compression sequential as by the time we find
the offsets for the next compression chunk, the last compression
chunk has finished.

As a test to fix this, we're going to try to add support for
course and fine grain split. The course split will just assign
big chunks to a bunch of fine grain split lambdas. The fine
grain split lambdas will handle all the actual offset logic.

In the next commit, we need to change the general iterator logic
to scan for the nearest separator token, as the course split does
not check for these.