benchmark: Switch last file to num files
We don't want to have to hard code the number of output files to
look for. To fix this, instead of tracking whether the current object
represents the last file, for every file, we indicate the total number
of files. This makes it easier to determine the total number of files
to look for as well as which file is the last file.