mzML: First iteration of generalizing mzML format
We want the mzML format to be like the other formats, where the split
function does not have to compute the actual offsets. Instead the
split functions assigns a big fix size chunk to a Lambda and that
Lambda adjusts the offsets to correct boundaries.

We still want to clean up the code a bit more, but we wrote some basic
tests to make sure the code generally works.