util: Clean up setup logic
Some of the variables and logic in the setup_lambda function was not
used anymore or there was a cleaner way to write the code. We remove
unused code and tried to make the rest of the code easier to
understand and easier to use with testing code. We still need to
update all Lambda functions to use the new code.