mzML: Account for chromatogram offsets
Using the footer offset to determine the end of spectra offset did
not work if the file contained chromatogram data. The amount of data
loaded if the file had chromatogram exceeded the amount of memory
that lambda can handle sometimes.

Instead we use the start of the chromatogram data (if it exists) to
determine the end of the spectra data.

We also make sure we track all offsets to make calculations easier.