setup: Map results to input / output variables
For some of the functions, we need results for earlier in the pipeline.
Before we were manually using the function index in the pipeline to
indicate which results to use. This isn't very good if we have to
change the pipeline. Instead we update the functions to take "input"
and "output" values to correctly set the prefix value.