scheduler: Another round of trying to improve
Okay I think this version actually makes sense. So we do have to
read every log entry unfortunately as this is the only way to
get the payloads to correctly trigger the child.

However, we change it so we only have to get each child once and
we don't have to reconstruct the log name. We only need to know the
prefix, bin id and file id.

Also modified tests.