scheduler: Fix simulation bugs
If we upload jobs, we accidentally retriggered the job since the queue
looks for items that start with prefix 1/. So we had to immediately add
the token job.

The re-execution causes the execution value to change sometimes so we
can't look for an exact log entry. We instead look for a prefix since
we know the bin values and file ID will not change, only the suffix.

We also had to change the lambda logic for determining if we should
run a function based on the execution value.

We had to add the execute logic to the upload script.