Re-execution: Fix bugs / hardcoding
Sometimes we need to re-execute a function and this changes
the execution parameter which affects the log we have to find.
We check for both execution = 0 and the original one now.

Also I have a bug related to finding logs. Until I figure out how
to do this correctly, I hardcoded each case.

Yay deadlines.