FASTQ format Iterator change
It looks like we need to use actual regex for FASTQ. To handle,
I add a regex parameter to the delimiter class. If regex is
specified, we use this to parse the array. After the deadline,
I'll try to generalize this, although I don't know how this will
work for mzML.