combine: Update logic to handle edge case
There's a potential edge case where we'll time out before all
files are present, but the given lambda represents the last
object and is therefore responsible for the combine. If we
terminate before all files appear, then we will write in the
logs that we have no payload, even though we should.

The reason this can occur is because sometimes there's lag
in AWS's API and so not all objects appear.

To prevent this case, if we "timeout", we won't write a log file
so this will be retriggered.