util: Update prior execution and log logic
Change prior execution to only be dependent on one log file. After
going back and forward, it seems like less effort and computation.
The downfall is we'll have less accurate measurements on total run-
time and lambdas invoked.

Also if we see the reexecute value and we've already executed the
command, reinvoke the children payloads.