setup: Move logic into class
We're in the process of porting Ripple to Google Functions. As
a results, we want setup.py to work with Google Functions as well.
The easiest thing seems to be to create a class to handle the
basic logic with abstract methods for the portion of the code
that needs to be AWS or Google specific.

All methods that need to be overwritten start with @abc.abstractmethod.

This commit is easier to view with -w.