profile: Basic profile implementation
This a very hard-coded first implementation to profile a ripple
application. This will be used to get stats of what split_size
should be used, number of lambdas, etc.

I still need to fix dependency problems related to imports in
statistics.

Also need to upload the results to S3, so we can use the results
to determine best parameters to use.